Adds configurable option to set message format
This allows the format of the messages produced by each formatter to be
configured by specifying a format string in the .pronto.yml config file.

Format can be overridden on a global basis by specifying the `format`
key at the top-level of the file, or it can be overridden on a
per-formatter basis by nesting `format` keys under the formatter's name
(as defined in `Formatter::FORMATTERS` constant). If a formatter does
not have a specific format defined in the config file, it will fall back
to the global format, and then fall back to the default format (which is
just `"%{msg}"`).

Format uses a Ruby format string and makes all of the attributes of
the `Message` object available for interpolation. The `TextFormatter`
additionally adds a special `TextMessageDecorator` to add colorized
and specially formatted values for file location and level. These were
in the original formatter code so I wanted to keep them.