add support for Ruby 3.0
With ruby 3.0 `rexml` is not installed by default. [It was moved to a
gem][1]. Adding the gem as a dependency fixed the specs on ruby 3.0.

Also switch TravisCI environment to bionic, do not require sudo.
On focal Ruby 2.7 can not be installed without additional
packages. With bionic CI is green from 2.3 up to 3.0.

[1]: https://www.ruby-lang.org/en/news/2020/12/25/ruby-3-0-0-released/